agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3>=1.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
